requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

# DB connectivity
psycopg2-binary==2.9.9
sqlite-utils==3.36.0

# Pydantic & typing
pydantic==2.6.3
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.6

# Workers / Background tasks (futuras mejoras)
celery==5.3.6
redis==5.0.3
